+++ async/delay.ts
@@ -12,6 +12,9 @@ export interface DelayOptions {
   persistent?: boolean;
 }
 
+// Make type available in browser environments; we catch the `ReferenceError` below
+declare const Deno: { unrefTimer(id: number): void };
+
 /**
  * Resolve a {@linkcode Promise} after a given amount of milliseconds.
  *
@@ -59,10 +62,10 @@ export function delay(ms: number, options: DelayOptions = {}): Promise<void> {
     signal?.addEventListener("abort", abort, { once: true });
     if (persistent === false) {
       try {
-        // @ts-ignore For browser compatibility
-        Deno.unrefTimer(i);
+        Deno.unrefTimer(+i);
       } catch (error) {
         if (!(error instanceof ReferenceError)) {
+          clearTimeout(+i);
           throw error;
         }
         // deno-lint-ignore no-console
@@ -86,8 +89,8 @@ function setArbitraryLengthTimeout(
   const queueTimeout = () => {
     currentDelay = delay - (Date.now() - start);
     timeoutId = currentDelay > I32_MAX
-      ? setTimeout(queueTimeout, I32_MAX)
-      : setTimeout(callback, currentDelay);
+      ? Number(setTimeout(queueTimeout, I32_MAX))
+      : Number(setTimeout(callback, currentDelay));
   };
 
   queueTimeout();