+++ http/cookie.ts
@@ -253,8 +253,8 @@ function validateDomain(domain: string) {
  */
 export function getCookies(headers: Headers): Record<string, string> {
   const cookie = headers.get("Cookie");
+  const out: Record<string, string> = Object.create(null);
   if (cookie !== null) {
-    const out: Record<string, string> = {};
     const c = cookie.split(";");
     for (const kv of c) {
       const [cookieKey, ...cookieVal] = kv.split("=");
@@ -264,9 +264,8 @@ export function getCookies(headers: Headers): Record<string, string> {
       const key = cookieKey.trim();
       out[key] = cookieVal.join("=");
     }
-    return out;
   }
-  return {};
+  return out;
 }
 
 /**