+++ yaml/stringify_test.ts
@@ -5,7 +5,10 @@
 
 import { assertEquals, assertThrows } from "@std/assert";
 import { stringify } from "./stringify.ts";
-import { stringify as unstableStringify } from "./unstable_stringify.ts";
+import {
+  type ImplicitType,
+  stringify as unstableStringify,
+} from "./unstable_stringify.ts";
 import { compare, parse } from "@std/semver";
 
 Deno.test({
@@ -864,6 +867,32 @@ Deno.test({
     assertEquals(
       unstableStringify(object, { quoteStyle: "'" }),
       `url: 'https://example.com'\n`,
+    );
+  },
+});
+
+Deno.test({
+  name: "unstableStringify() handles custom types",
+  fn() {
+    const foo: ImplicitType = {
+      tag: "tag:custom:smile",
+      resolve: (data: string): boolean => data === "=)",
+      construct: (): string => "🙂",
+      predicate: (data: unknown): data is string => data === "🙂",
+      kind: "scalar",
+      represent: (): string => "=)",
+    };
+
+    assertEquals(
+      unstableStringify({
+        title: "🙂",
+        tags: ["🙂", "bar"],
+      }, { extraTypes: [foo] }),
+      `title: =)
+tags:
+  - =)
+  - bar
+`,
     );
   },
 });