+++ Releases.md
@@ -1,3 +1,154 @@
+### 2025.11.17
+
+#### @std/assert 1.0.16 (patch)
+
+- fix(assert,streams): handle SharedArrayBuffer being disabled in some browser
+  context (#6856)
+
+#### @std/cache 0.2.1 (patch)
+
+- feat(cache): add cleanup callback to `LruCache` and `TtlCache` (#6857)
+
+#### @std/cbor 0.1.9 (patch)
+
+- docs(cbor): remove h2 and change spelling from UK -> US (#6847)
+
+#### @std/cli 1.0.24 (patch)
+
+- feat(cli/unstable): custom progress bar refresh interval (#6862)
+- fix(cli/unstable): typo in ProgressBar example (#6848)
+
+#### @std/fs 1.0.20 (patch)
+
+- fix(fs): "Deno is not defined" when using the module in browser (#6868)
+- fix(fs): `expandGlob`/`expandGlobSync` - match non-glob path segments
+  containing escaped glob chars (#6788)
+- fix(fs,path): `expandGlob`, `expandGlobSync`, and `isGlob` match
+  `@(alternate)` syntax with only one member (#6850)
+
+#### @std/http 1.0.22 (patch)
+
+- feat(http/unstable): add support for extra headers in serveFile (#6842)
+- fix(http): runtime types for `getCookies` (#6853)
+
+#### @std/math 0.0.0 (patch)
+
+- feat(math/unstable): add `math` package with basic math utilities (#6823)
+
+#### @std/path 1.1.3 (patch)
+
+- fix(fs,path): `expandGlob`, `expandGlobSync`, and `isGlob` match
+  `@(alternate)` syntax with only one member (#6850)
+
+#### @std/random 0.1.4 (patch)
+
+- docs(random): use PRNG initialization with secure random seed (#6855)
+
+#### @std/semver 1.0.7 (patch)
+
+- refactor(semver): reduce nesting in `testComparatorSet()` (#6866)
+
+#### @std/streams 1.0.14 (patch)
+
+- fix(assert,streams): handle SharedArrayBuffer being disabled in some browser
+  context (#6856)
+
+### 2025.10.07a
+
+#### @std/internal 1.0.12 (patch)
+
+- fix(internal): fix error in module reference (#6845)
+
+### 2025.10.07
+
+#### @std/assert 1.0.15 (patch)
+
+- feat(assert/unstable): truncate big diffs (#6835)
+- fix(assert): add support for checking ArrayBuffer equality (#6821)
+
+#### @std/async 1.0.15 (patch)
+
+- feat(async/unstable): add `ensureLastCall` option for `throttle` (#6800)
+- fix(async/unstable): fix handling of async predicate in waitFor (#6839)
+- fix(async): convert setTimeout return values to Number (fixing type error when
+  used with Node.js) (#6833)
+- fix(async): pass numeric value of `delay` timer to `unrefTimer` when
+  `persistent=false` (#6822)
+
+#### @std/cli 1.0.23 (patch)
+
+- fix(cli): improve Node compatibility by removing top-level Deno API usage
+  (#6837)
+- fix(cli): ensure unknown() is not called for collect args (parseArgs) (#6813)
+
+#### @std/html 1.0.5 (patch)
+
+- feat(html/unstable): add escapeJs and escapeCss functions (#6782)
+
+#### @std/http 1.0.21 (patch)
+
+- fix(http,random,streams,tar,webgpu): update types for Deno 2.5 (#6817)
+
+#### @std/internal 1.0.11 (patch)
+
+- chore(internal): support truncating of big diffs (#6835)
+
+#### @std/random 0.1.3 (patch)
+
+- fix(http,random,streams,tar,webgpu): update types for Deno 2.5 (#6817)
+
+#### @std/semver 1.0.6 (patch)
+
+- fix(semver): allow users to set prerelease number during increment (#6825)
+
+#### @std/streams 1.0.13 (patch)
+
+- fix(http,random,streams,tar,webgpu): update types for Deno 2.5 (#6817)
+
+#### @std/tar 0.1.9 (patch)
+
+- fix(http,random,streams,tar,webgpu): update types for Deno 2.5 (#6817)
+
+#### @std/testing 1.0.16 (patch)
+
+- feat(testing/unstable): add `it.todo` `test.todo` and `describe.todo` API
+  (#6712)
+
+#### @std/toml 1.0.11 (patch)
+
+- test(toml): test with the official toml-test suite (#6798)
+
+#### @std/webgpu 0.224.9 (patch)
+
+- fix(http,random,streams,tar,webgpu): update types for Deno 2.5 (#6817)
+
+#### @std/yaml 1.0.10 (patch)
+
+- feat(yaml/unstable): allow to add custom types for parse and stringify (#6841)
+
+### 2025.09.04
+
+#### @std/cli 1.0.22 (patch)
+
+- fix(cli/unstable): update typing of ProgressBarStream for TS 5.9 (#6811)
+
+#### @std/net 1.0.6 (patch)
+
+- feat(net/unstable): add matchSubnets on @std/net (#6786)
+
+#### @std/streams 1.0.12 (patch)
+
+- deprecation(streams/unstable): AbortStream (#6799)
+
+#### @std/tar 0.1.8 (patch)
+
+- refactor(tar): clean up TarStream (#6783)
+- test(tar): improve TarStream's tests (#6793)
+
+#### @std/toml 1.0.10 (patch)
+
+- fix(toml): nested table arrays (#6794)
+
 ### 2025.08.13
 
 #### @std/assert 1.0.14 (patch)